global-exception & validation

+++ src/main/java/com/thebinh/identity_service/controller/UserController.java
@@ -1,7 +1,10 @@
 package com.thebinh.identity_service.controller;
 
 import com.thebinh.identity_service.domain.User;
+import com.thebinh.identity_service.exception.IdInvalidException;
+import com.thebinh.identity_service.exception.UsernameInvalidException;
 import com.thebinh.identity_service.service.UserService;
+import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,14 +34,25 @@ public class UserController {
 
     //create a user
     @PostMapping("/user")
-    public String createUser(@RequestBody User user){
+    public String createUser(@Valid @RequestBody User user) throws UsernameInvalidException{
+        User checkExitsUser = this.userService.handleUserByUsername(user.getUsername());
+        if(checkExitsUser != null){
+            throw new UsernameInvalidException("username đã tồn tại");
+        }
         this.userService.handleSaveUser(user);
         return "đã tạo người dùng ở trong DB";
     }
 
     //delete user
     @DeleteMapping("/user/{id}")
-    public String deleteUser(@PathVariable("id") long id){
+    public String deleteUser(@PathVariable("id") long id) throws IdInvalidException {
+        if(id >= 1500){
+            throw new IdInvalidException("id không được vượt quá 1500");
+        }
+        User user = this.userService.handleUserById(id);
+        if(user == null){
+            throw new IdInvalidException("người dùng không tồn tại trong hệ thống");
+        }
         this.userService.deleteUserById(id);
         return "đã xóa người dùng trong db";
     }
